Basic Crud App/DB connection
The basic setup for the creation of notes, as well as time stamps and subtitle have all been complete. As well as a subpage for the /notes. this allows for separation from the notes from the rest of the app. Edit and delete buttons are added but not yet implemented. Just a basic set up with proper local host connection, MongoDB connection and proper npm setup.