Fully functioning app
The application now fully acts as it should, the edit, delete and update buttons are now all working. This also includes the image upload capabilities from Multer. So in conclusion all functionalities work and now it ready to start inplementing front end visuals to make the website cleaner and a more user friendly applicaiton.